refactor(dot11): extract Dot11FrameType into frame_type module
