+++ crafter/src/protocols/link/dot11/mod.rs
@@ -14,9 +14,11 @@ use crate::registry::ProtocolRegistry;
 use super::append_llc_snap_packet_with_registry;
 
 mod constants;
+mod frame_type;
 mod util;
 
 pub use self::constants::*;
+pub use self::frame_type::Dot11FrameType;
 use self::util::{dot11_hex_bytes, set_flag, set_subfield, value_or_copy};
 
 /// Raw IEEE 802.11 management tagged parameter.
@@ -1805,63 +1807,6 @@ fn decode_dot11_tagged_parameters(bytes: &[u8]) -> Result<Vec<Dot11TaggedParamet
     }
 
     Ok(tags)
-}
-
-/// IEEE 802.11 frame type subfield.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[non_exhaustive]
-pub enum Dot11FrameType {
-    /// Management frame.
-    Management,
-    /// Control frame.
-    Control,
-    /// Data frame.
-    Data,
-    /// Extension frame.
-    Extension,
-    /// Unknown caller-supplied value preserved verbatim.
-    Unknown(u8),
-}
-
-impl Dot11FrameType {
-    /// Create a frame type from its raw numeric value.
-    pub const fn from_raw(value: u8) -> Self {
-        match value {
-            DOT11_FRAME_TYPE_MANAGEMENT => Self::Management,
-            DOT11_FRAME_TYPE_CONTROL => Self::Control,
-            DOT11_FRAME_TYPE_DATA => Self::Data,
-            DOT11_FRAME_TYPE_EXTENSION => Self::Extension,
-            value => Self::Unknown(value),
-        }
-    }
-
-    /// Raw numeric value.
-    pub const fn raw(self) -> u8 {
-        match self {
-            Self::Management => DOT11_FRAME_TYPE_MANAGEMENT,
-            Self::Control => DOT11_FRAME_TYPE_CONTROL,
-            Self::Data => DOT11_FRAME_TYPE_DATA,
-            Self::Extension => DOT11_FRAME_TYPE_EXTENSION,
-            Self::Unknown(value) => value,
-        }
-    }
-
-    /// Short stable label.
-    pub fn label(self) -> String {
-        dot11_frame_type_label(self.raw())
-    }
-}
-
-impl From<u8> for Dot11FrameType {
-    fn from(value: u8) -> Self {
-        Self::from_raw(value)
-    }
-}
-
-impl From<Dot11FrameType> for u8 {
-    fn from(value: Dot11FrameType) -> Self {
-        value.raw()
-    }
 }
 
 /// IEEE 802.11 management frame subtype subfield.
